set siginterrupt flag to False
This patch sets the siginterrupt flag to False so that interrupted system calls
are automatically restarted. This notably fixes applications using the Precice
coupling library, which does not currently handle EINTR read errors. This patch
also changes the unit tests to SIGINT-based control, so that signal handlers
are not blocked while waiting for user input.